jarvis_ai/requirements.txt
SpeechRecognition
pyttsx3
pyaudio
google-generativeai
openai
python-dotenv
elevenlabs
edge-tts
pygame
# TTS - Disabled due to build issues
duckduckgo-search
spotipy
psutil
pyperclip
requests
pvporcupine
pvcheetah
